backend/requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
llama-parse
chromadb
llama-index-vector-stores-chroma
nest_asyncio
openai
pypdf